+++ StripePaymentSheet/StripePaymentSheet/Source/PaymentSheet/PaymentSheet+DeferredAPI.swift
@@ -48,6 +48,13 @@ extension PaymentSheet {
                 switch intentConfig.mode {
                 case .payment:
                     let paymentIntent = try await configuration.apiClient.retrievePaymentIntent(clientSecret: clientSecret, expand: ["payment_method"])
+                    let completion = { (status: STPPaymentHandlerActionStatus, paymentIntent: STPPaymentIntent?, error: NSError?, deferredIntentConfirmationType: STPAnalyticsClient.DeferredIntentConfirmationType) in
+                        if let paymentIntent {
+                            setDefaultPaymentMethodIfNecessary(actionStatus: status, intent: .paymentIntent(paymentIntent), configuration: configuration)
+                        }
+                        completion(makePaymentSheetResult(for: status, error: error), deferredIntentConfirmationType)
+                    }
+
                     // Check if it needs confirmation
                     if [STPPaymentIntentStatus.requiresPaymentMethod, STPPaymentIntentStatus.requiresConfirmation].contains(paymentIntent.status) {
                         // 4a. Client-side confirmation
@@ -70,8 +77,8 @@ extension PaymentSheet {
                         paymentHandler.confirmPayment(
                             paymentIntentParams,
                             with: authenticationContext
-                        ) { status, _, error in
-                            completion(makePaymentSheetResult(for: status, error: error), .client)
+                        ) { status, paymentIntent, error in
+                            completion(status, paymentIntent, error, .client)
                         }
                     } else {
                         // 4b. Server-side confirmation
@@ -79,12 +86,18 @@ extension PaymentSheet {
                             for: paymentIntent,
                             with: authenticationContext,
                             returnURL: configuration.returnURL
-                        ) { status, _, error in
-                            completion(makePaymentSheetResult(for: status, error: error), .server)
+                        ) { status, paymentIntent, error in
+                            completion(status, paymentIntent, error, .server)
                         }
                     }
                 case .setup:
                     let setupIntent = try await configuration.apiClient.retrieveSetupIntent(clientSecret: clientSecret, expand: ["payment_method"])
+                    let completion = { (status: STPPaymentHandlerActionStatus, setupIntent: STPSetupIntent?, error: NSError?, deferredIntentConfirmationType: STPAnalyticsClient.DeferredIntentConfirmationType) in
+                        if let setupIntent {
+                            setDefaultPaymentMethodIfNecessary(actionStatus: status, intent: .setupIntent(setupIntent), configuration: configuration)
+                        }
+                        completion(makePaymentSheetResult(for: status, error: error), deferredIntentConfirmationType)
+                    }
                     if [STPSetupIntentStatus.requiresPaymentMethod, STPSetupIntentStatus.requiresConfirmation].contains(setupIntent.status) {
                         // 4a. Client-side confirmation
                         try PaymentSheetDeferredValidator.validate(setupIntent: setupIntent, intentConfiguration: intentConfig)
@@ -97,8 +110,8 @@ extension PaymentSheet {
                         paymentHandler.confirmSetupIntent(
                             setupIntentParams,
                             with: authenticationContext
-                        ) { status, _, error in
-                            completion(makePaymentSheetResult(for: status, error: error), .client)
+                        ) { status, setupIntent, error in
+                            completion(status, setupIntent, error, .client)
                         }
                     } else {
                         // 4b. Server-side confirmation
@@ -106,8 +119,8 @@ extension PaymentSheet {
                             for: setupIntent,
                             with: authenticationContext,
                             returnURL: configuration.returnURL
-                        ) { status, _, error in
-                            completion(makePaymentSheetResult(for: status, error: error), .server)
+                        ) { status, setupIntent, error in
+                            completion(status, setupIntent, error, .server)
                         }
                     }
                 }